Add files via upload
This is the supplementary material of the paper "Seamounts Ecological Modelling: A Comprehensive Review and Assessment of Modelling Suitability to Emergent Challenges". 

Supplementary material 1 (SM1) presents the output of the systematic review of ecological models applied to seamounts ecosystems. Sheet 1 presents the full description of abbreviations used in Sheet2.

Supplementary material 2 (SM2) presents a very brief description of the models found during the systematic review of ecological models applied to seamounts ecosystems. We present a short description of Model categories (i.e., broader aim of the model), followed by Types of Models (i.e., type of framework/software).